requirements.txt
selenium
pytest
Pillow
allure-pytest
